refactor: split Intermediate Representation models and SDG models into subpackages
