+++ src/main/java/edu/university/ecs/lab/common/models/ir/Flow.java
@@ -1,4 +1,4 @@
-package edu.university.ecs.lab.common.models;
+package edu.university.ecs.lab.common.models.ir;
 
 import com.google.gson.JsonNull;
 import com.google.gson.JsonObject;
