+++ src/main/java/edu/university/ecs/lab/common/models/sdg/NetworkGraph.java
@@ -1,7 +1,8 @@
-package edu.university.ecs.lab.common.models;
+package edu.university.ecs.lab.common.models.sdg;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
+import edu.university.ecs.lab.common.models.ir.*;
 import edu.university.ecs.lab.common.models.serialization.JsonSerializable;
 import lombok.AllArgsConstructor;
 import lombok.Data;
