+++ src/main/java/edu/university/ecs/lab/common/models/serialization/package-info.java
@@ -6,11 +6,11 @@
  * - {@link edu.university.ecs.lab.common.models.serialization.JsonSerializable}: Interface for classes
  *   that can be serialized to JSON objects.
  * - {@link edu.university.ecs.lab.common.models.serialization.MethodCallDeserializer}: Deserializer
- *   for converting JSON to {@link edu.university.ecs.lab.common.models.MethodCall} and
- *   {@link edu.university.ecs.lab.common.models.RestCall} objects.
+ *   for converting JSON to {@link edu.university.ecs.lab.common.models.ir.MethodCall} and
+ *   {@link edu.university.ecs.lab.common.models.ir.RestCall} objects.
  * - {@link edu.university.ecs.lab.common.models.serialization.MethodDeserializer}: Deserializer for
- *   converting JSON to {@link edu.university.ecs.lab.common.models.Method} and
- *   {@link edu.university.ecs.lab.common.models.Endpoint} objects.
+ *   converting JSON to {@link edu.university.ecs.lab.common.models.ir.Method} and
+ *   {@link edu.university.ecs.lab.common.models.ir.Endpoint} objects.
  * <p>
  * These classes facilitate conversion between Java objects and JSON representations.
  */
