+++ src/main/java/edu/university/ecs/lab/common/utils/SourceToObjectUtils.java
@@ -9,9 +9,9 @@ import com.github.javaparser.ast.expr.*;
 import com.github.javaparser.ast.nodeTypes.NodeWithSimpleName;
 import edu.university.ecs.lab.common.config.Config;
 import edu.university.ecs.lab.common.error.Error;
-import edu.university.ecs.lab.common.models.*;
 import edu.university.ecs.lab.common.models.enums.ClassRole;
 import edu.university.ecs.lab.common.models.enums.HttpMethod;
+import edu.university.ecs.lab.common.models.ir.*;
 import edu.university.ecs.lab.intermediate.utils.StringParserUtils;
 
 import java.io.File;
