+++ src/main/java/edu/university/ecs/lab/detection/antipatterns/services/WobblyServiceInteractionService.java
@@ -1,10 +1,10 @@
 package edu.university.ecs.lab.detection.antipatterns.services;
 
-import edu.university.ecs.lab.common.models.Annotation;
-import edu.university.ecs.lab.common.models.JClass;
-import edu.university.ecs.lab.common.models.Method;
-import edu.university.ecs.lab.common.models.Microservice;
-import edu.university.ecs.lab.common.models.MicroserviceSystem;
+import edu.university.ecs.lab.common.models.ir.Annotation;
+import edu.university.ecs.lab.common.models.ir.JClass;
+import edu.university.ecs.lab.common.models.ir.Method;
+import edu.university.ecs.lab.common.models.ir.Microservice;
+import edu.university.ecs.lab.common.models.ir.MicroserviceSystem;
 import edu.university.ecs.lab.detection.antipatterns.models.WobblyServiceInteraction;
 
 import java.util.ArrayList;
