+++ src/main/java/edu/university/ecs/lab/detection/architecture/models/UseCase3.java
@@ -1,13 +1,13 @@
 package edu.university.ecs.lab.detection.architecture.models;
 
-import edu.university.ecs.lab.common.models.Endpoint;
-import edu.university.ecs.lab.common.models.JClass;
-import edu.university.ecs.lab.common.models.Microservice;
-import edu.university.ecs.lab.common.models.MicroserviceSystem;
-import edu.university.ecs.lab.common.models.RestCall;
 import edu.university.ecs.lab.common.models.enums.ClassRole;
 import edu.university.ecs.lab.delta.models.Delta;
 import edu.university.ecs.lab.delta.models.enums.ChangeType;
+import edu.university.ecs.lab.common.models.ir.Endpoint;
+import edu.university.ecs.lab.common.models.ir.JClass;
+import edu.university.ecs.lab.common.models.ir.Microservice;
+import edu.university.ecs.lab.common.models.ir.MicroserviceSystem;
+import edu.university.ecs.lab.common.models.ir.RestCall;
 import edu.university.ecs.lab.detection.architecture.models.enums.Scope;
 import lombok.Data;
 
@@ -60,13 +60,17 @@ public class UseCase3 extends UseCase {
         return metaData;
     }
 
-    public static List<UseCase3> scan(Delta delta, MicroserviceSystem microserviceSystem) {
-        List<UseCase3> useCases = new ArrayList<>();
-        
-        if (!delta.getChangeType().equals(ChangeType.ADD) || !delta.getClassChange().getClassRole().equals(ClassRole.SERVICE)) {
-            return useCases;
+    public static UseCase3 scan(RestCall restCall, MicroserviceSystem microserviceSystem){
+        for (Microservice microservice : microserviceSystem.getMicroservices()){
+            for(JClass controller : microservice.getControllers()){
+                for (Endpoint endpoint : controller.getEndpoints()){
+                    if (RestCall.matchEndpoint(restCall, endpoint)){
+                        return null;
+                    }
+                }
+            }
         }
-        
+
         for (RestCall restCall : delta.getClassChange().getRestCalls()) {
             if (!existsInSystem(restCall, microserviceSystem)) {
                 UseCase3 useCase3 = new UseCase3();
@@ -76,7 +80,7 @@ public class UseCase3 extends UseCase {
                 useCases.add(useCase3);
             }
         }
-        
+
         return useCases;
     }
